refactor: extraire exporterEnCsv dans un hook dédié pour ComparaisonTerritoiresIndicateur
